+++ R/geom-column.r
@@ -96,7 +96,7 @@ hc_column <- function(chart, spec, opts, geom_params, use_js = TRUE, ...) {
 #' hd_make(spec_col, "column", opts_col)
 #'
 #' # Static ggplot2
-#' hd_make(spec_col, "column", opts_col, backend = "ggplot2")
+#' hd_make(spec_col, "column", opts_col, backend = "static")
 #'
 #' # Composable style
 #' p <- hd(survey, x = "age_group", y = "pct", group = "kjonn")
@@ -108,7 +108,7 @@ hc_column <- function(chart, spec, opts, geom_params, use_js = TRUE, ...) {
 #' # Pass an existing hd_spec
 #' spec <- hd_spec(survey, x = "age_group", y = "pct", group = "kjonn", n = "n")
 #'
-#' hd(spec, backend = "ggplot2") +
+#' hd(spec, backend = "static") +
 #'  hd_geom_column() +
 #'  hd_opts(title = "Health survey", ylim = c(0, 80))
 #'